backend/requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.32.1
streamlit==1.41.1
streamlit-chat==0.1.1
edge-tts==6.1.19
python-multipart==0.0.20
python-dotenv==1.0.1
PyPDF2==3.0.1
numpy==1.26.4
opencv-python-headless==4.10.0.84
gTTS==2.5.4
Pillow==10.4.0
langchain==0.3.27
langchain-community==0.3.27
langchain-core==0.3.81
langchain-text-splitters==0.3.9
langchain-cohere==0.4.6
faiss-cpu==1.9.0
